Refactor gateway namespaces in preparation to #90
