+++ lib/Service/Gateway/Telegram/Gateway.php
@@ -4,8 +4,9 @@ declare(strict_types=1);
 
 /**
  * @author Christoph Wurst <christoph@winzerhof-wurst.at>
+ * @author André Fondse <andre@hetnetwerk.org>
  *
- * Nextcloud - Two-factor Gateway
+ * Nextcloud - Two-factor Gateway for Telegram
  *
  * This code is free software: you can redistribute it and/or modify
  * it under the terms of the GNU Affero General Public License, version 3,
@@ -21,18 +22,20 @@ declare(strict_types=1);
  *
  */
 
-namespace OCA\TwoFactorGateway\Service\Gateway;
+namespace OCA\TwoFactorGateway\Service\Gateway\Telegram;
 
 use Exception;
 use OCA\TwoFactorGateway\Exception\SmsTransmissionException;
-use OCA\TwoFactorGateway\Service\IGateway;
+use OCA\TwoFactorGateway\Service\Gateway\IGateway;
 use OCP\Http\Client\IClient;
 use OCP\Http\Client\IClientService;
 use OCP\IConfig;
 use OCP\IL10N;
 use OCP\IUser;
+use Telegram\Bot\Api;
+use Telegram\Bot\Objects\Update;
 
-class WebSmsGateway implements IGateway {
+class Gateway implements IGateway {
 
 	/** @var IClient */
 	private $client;
@@ -56,26 +59,42 @@ class WebSmsGateway implements IGateway {
 	 * @param string $idenfier
 	 * @param string $message
 	 *
-	 * @throws SmsTransmissionException
+	 * @throws \Telegram\Bot\Exceptions\TelegramSDKException
 	 */
 	public function send(IUser $user, string $idenfier, string $message) {
-		$user = $this->config->getAppValue('twofactor_gateway', 'websms_de_user');
-		$password = $this->config->getAppValue('twofactor_gateway', 'websms_de_password');
-		try {
-			$this->client->post('https://api.websms.com/rest/smsmessaging/text', [
-				'headers' => [
-					'Authorization' => 'Basic ' . base64_encode("$user:$password"),
-					'Content-Type' => 'application/json',
-				],
-				'json' => [
-					'messageContent' => $message,
-					'test' => false,
-					'recipientAddressList' => [$idenfier],
-				],
-			]);
-		} catch (Exception $ex) {
-			throw new SmsTransmissionException();
+		$token = $this->config->getAppValue('twofactor_gateway', 'telegram_bot_token', null);
+		// TODO: token missing handling
+
+		$api = new Api($token);
+		$chatId = $this->getChatId($user, $api, (int)$idenfier);
+
+		$api->sendMessage([
+			'chat_id' => $chatId,
+			'text' => $message,
+		]);
+	}
+
+	private function getChatId(IUser $user, Api $api, int $userId): int {
+		$chatId = $this->config->getUserValue($user->getUID(), 'twofactor_gateway', 'telegram_chat_id', null);
+
+		if (!is_null($chatId)) {
+			return (int)$chatId;
 		}
+
+		$updates = $api->getUpdates();
+		/** @var Update $update */
+		$update = current(array_filter($updates, function (Update $data) use ($userId) {
+			if ($data->message->text === "/start" && $data->message->from->id === $userId) {
+				return true;
+			}
+			return false;
+		}));
+		// TODO: handle missing `/start` message and `$update` null values
+
+		$chatId = $update->message->chat->id;
+		$this->config->setUserValue($user->getUID(), 'twofactor_gateway', 'chat_id', $chatId);
+
+		return (int)$chatId;
 	}
 
 	/**
@@ -85,13 +104,13 @@ class WebSmsGateway implements IGateway {
 	 * @return string
 	 */
 	public function getShortName(): string {
-		return 'SMS';
+		return 'Telegram';
 	}
 
 	/**
 	 * @return string
 	 */
 	public function getProviderDescription(): string {
-		return $this->l10n->t('Authenticate via SMS');
+		return $this->l10n->t('Authenticate via Telegram');
 	}
 }
